pyproject.toml
[tool.poetry]
name = "sarafan"
version = "0.1.0"
description = ""
authors = ["zaritskiyaa <zaritskiyaa@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
pillow = "^10.4.0"
numpy = "^2.0.1"
pyjwt = "^2.9.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
